Massive changes to config file handling
