+++ src/configfile.h
@@ -1,0 +1,32 @@
+#ifndef CONFIG_FILE_H__
+#define CONFIG_FILE_H__
+
+#include <string>
+#include <map>
+#include <vector>
+#include <boost/algorithm/string/split.hpp>
+#include <boost/algorithm/string/classification.hpp>
+
+#include "mlink.h"
+#include "serial.h"
+#include "asyncsocket.h"
+
+class ConfigFile
+{
+    std::map   <std::string, std::string> content_;
+    std::vector<std::string>              sections_;
+public:
+    ConfigFile();
+    ConfigFile(std::string const& configFile);
+
+    std::vector<std::string> GetSections();
+
+    bool boolValue(std::string const& section, std::string const& entry, bool* value);
+    bool intValue(std::string const& section, std::string const& entry, int* value);
+    bool strValue(std::string const& section, std::string const& entry, std::string* value);
+};
+
+void readLinkInfo(ConfigFile* _configFile, std::string thisSection, link_info* _info);
+int readConfigFile(std::string &filename, std::vector<std::shared_ptr<mlink> > &links);
+
+#endif
